Moved isProgramManagerForEvent into user file from the volunteer file. moved its test over to the user test as well.
